refactor: type 패키지 도메인 패키지로 이동
