+++ src/main/java/lotto/domain/type/ExceptionType.java
@@ -1,4 +1,4 @@
-package lotto.type;
+package lotto.domain.type;
 
 import lombok.Getter;
