+++ src/test/java/lotto/domain/LottoCalculatorTest.java
@@ -1,7 +1,7 @@
 package lotto.domain;
 
 import lotto.LottoCalculatorTestHelper;
-import lotto.type.RankType;
+import lotto.domain.type.RankType;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
